Iteration4_Session3 - unique and all word counting

+++ src/test/java/com/george/interview/counter/WordCounterTest.java
@@ -1,12 +1,9 @@
 package com.george.interview.counter;
 
-import kotlin.Pair;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 
 import static com.george.interview.counter.WordCounterTestSuites.*;
 
@@ -18,7 +15,7 @@ public class WordCounterTest {
         testSuite.forEach(pair ->
                 Assertions.assertEquals(
                         pair.getSecond().longValue(),
-                        counter.count(pair.getFirst(), Collections.emptyList()).longValue()
+                        counter.count(pair.getFirst(), Collections.emptyList())
                 )
         );
     }
@@ -29,7 +26,7 @@ public class WordCounterTest {
         testSuiteBadWords.forEach(pair ->
                 Assertions.assertEquals(
                         pair.getSecond().longValue(),
-                        counter.count(pair.getFirst(), excludedWords).longValue()
+                        counter.count(pair.getFirst(), excludedWords)
                 )
         );
     }
